Fits Semiparametric Regression Models for Panel Count Survival Data

Fits an proportional means model: $$\Lambda (t; X_i) = E[N_i(t)|X_i] = \Lambda(t)e^{X_i ' \beta},$$ where \(\beta\) is a \(p \times 1\) vector of covariate coefficient and \(\Lambda(\cdot)\) is a completely unspecified baseline mean function.

Estimating procedures include: Wang-Yan's augmented estimating equations ("AEE", "AEEX"), Huang-Wang-Zhang's method ("HWZ"), Zhang's maximum pseudo-likelihood ("MPL"), Maximum pseudolikelihood with I-Splines ("MPLs"), Maximum likelihood with I-Splines ("MLs"), Sun-Wei's method ("EE.SWa", "EE.SWb", "EE.SWc"), and Hu-Sun-Wei's method ("EE.HSWc", "EE.HSWm").

The function can also fits an accelerated mean model ("AMM"): $$\Lambda (t; X_i) = E[N_i(t)|X_i] = \Lambda(te^{X_i ' \beta})e^{X_i ' \beta}.$$

panelReg(formula, data, method = c("AEE", "AEEX", "HWZ", "MPL", "MPLs", "MLs",
  "AMM", "EE.SWa", "EE.SWb", "EE.SWc", "EE.HSWc", "EE.HSWm"), se = c("NULL",
  "smBootstrap", "Bootstrap", "Impute", "Sandwich"), control = list())

Arguments

formula
A formula object, with the response on the left of a "~" operator, and the terms on the right. The response must be a panel count survival object as returned by function PanelSurv.

data
A data.frame in which to interpret the variables named in the formula. Three variables including subjects' id, observation times, and number of new events since last observation time are required to feed into function PanelSurv as response. At least one covariate variable is required.

method
Fitting method. See ‘Details’.

se
Standard error estimation method. See ‘Details’.

control
A list of control parameters. See ‘Details’.

Value

An object of S3 class "panelReg" representing the fit. See panelReg.object for details.

Details

Some assumptions details about the observation times and censoring time need clarification. Three possible scenarios of observation times are considered: 1) independent observation times -- the observation times are independent of the underlying recurrent event process; 2) conditional independent observation times -- the observation times are independent of the event process given observed covariates; 3) informative observation times -- after conditioning on observed covariates, the observation times and the event process are still dependent through an unobserved multiplicative frailty. Similarly, the three scenarios apply to the censoring time.

"AEE" and "AEEX" are the augmented estimating equation methods under conditional independent censoring and informative censoring respectively. Both allow informative observation times.

"HWZ" is Huang-Wang-Zhang's method. It allows both information observation times and informative censoring time. It does not need to specify the dependence structure or model the frailty.

"MPL" and "MPLs" are maximum pseudolikelihood methods, with nonparametric and monotone spline estimates of the baseline mean function respectively. They assume conditional independent observation times and censoring time. The underlying event process is assumed to be Poisson, and the within subjects dependence is ignored.

"MLs" is maximum likelihood method with monotone splines estimates of the baseline mean function. It assumes conditional independent observation times and censoring time, and a Poisson underlying event process.

"EE.SWa", "EE.SWb" and "EE.SWc" are estimating equation approaches based on Sun-Wei's methods. The first assumes independent observation times and censoring time. The second assumes conditional independent observation times but independent censoring time. The third assumes conditional independent observation times and censoring time. All three variations work on centered covariates and avoid estimating the baseline mean.

"EE.HSWc" and "EE.HSWm" are estimating equation approaches based on Hu-Sun-Wei's methods. The first assumes independent observation times and censoring time. The second assumes conditional independent observation times but independent censoring time. Both variations work on centered covariates and avoid estimating the baseline mean.

"AMM" is the accelerated mean model. The observation time process is allowed to be correlated with the underlying recurrent event process through a frailty variable, which does not need to be specified. The model also allows marginal interpretations for the regression parameters and connects naturally with AFT model. See Chiou et al. (2017) for more details.

For standard errors estimation method:

"NULL" means do not calculate standard errors; "smBootstrap" is the smoothed bootstrap estimation method that works with "AMM". "Bootstrap" works with all fitting methods; "Impute" is the multiple imputation method that works with "AEE" and "AEEX"; "Sandwich" is the robust sandwich estimation method that works with "AEE" and "AEEX".

The control argument is a list that can supply any of the following components:

betaInit:

Object of class "numeric", initial value for covariate coefficient, default 0.

interval:

Object of class "numeric", initial search interval for solving beta, default (-5, 5).

maxIter:

Object of class "numeric", maximum iteration allowed, default 500 for "AEEX" and "HWZ", 150 for others.

absTol:

Object of class "numeric", absolute tolerance, default 1e-6.

relTol:

Object of class "numeric", relative tolerance, default 1e-6.

a:

Object of class "numeric", a tune parameter, default 0.1. In the case of gamma frailty, "a" corresponds to the value of both shape and rate parameters.

R:

Object of class "numeric", number of bootstrap or imputation, default 30.

References

Chiou, S., Xu, G., Yan, J., and Huang, C.-Y. (2017). Semiparametric estimation of the accelerated mean model with panel count data under informative examination times. Biometrics, to appear. <doi: 10.1111/biom.12840>.

Huang, C.-Y., Wang, M., and Zhang, Y. (2006). Analysing panel count data with informative observation times. Biometrika, 93(4), 763--776.

Hu, X. J., Sun, J. and Wei, L. J. (2003). Regression parameter estimation from panel counts. Scandinavian Journal of Statistics, 30, 25--43.

Lu, M., Zhang, Y., and Huang, J. (2007). Estimation of the mean function with panel count data using monotone polynomial splines. Biometrika, 94(3), 705--718.

Sun, J. and Wei, L. J. (2000). Regression analysis of panel count data with covariates-dependent observation and censoring times. Journal of the Royal Statistical Society, Series B: Statistical Methodology, 62(2), 293--302.

Wang, X. and Yan, J. (2011). Fitting semiparametric regressions for panel count survival data with an R package spef. Computer Methods and Programs in Biomedicine, 104(2), 278--285.

Wang, X. and Yan, J. (2013). Augmented estimating equations for semiparametric panel count regression with informative observation times and censoring time. Statistica Sinica, 359--381.

Zhang, Y. (2002). A Semiparametric pseudolikelihood estimation method for panel count data. Biometrika, 89(1), 39--48.

See also

Examples

# NOT RUN {
data(blaTum)
## Fit the bladder tumor data set
formula <- PanelSurv(id, time, count) ~ num + size + treatment

panelReg(formula, data = blaTum, method = "AEE", se = "Sandwich")
panelReg(formula, data = blaTum, method = "AEEX", se = "Impute",
         control = list(a = 0.1, R = 30))
panelReg(formula, data = blaTum, method = "HWZ", se = "Bootstrap",
         control = list(R = 30))
panelReg(formula, data = blaTum, method = "MLs", se = "NULL")
panelReg(formula, data = blaTum, method = "EE.SWa", se = "Bootstrap",
         control = list(R = 30))
panelReg(formula, data = blaTum, method = "EE.HSWc", se = "Bootstrap",
         control = list(R = 30))
# }