Semiparametric Panel Count Regression Object

This S3 class of objects is returned by the panelReg class of functions to represent a fitted semiparametric panel count regression model. Objects of this class have methods for the functions print and plot.

Value

beta

a vector, estimated coefficients of the linear predictor.

baseline

a step function or spline function, estimated cumulative baseline mean.

timeGrid

a vector, ordered unique set of observation times.

lambda

a vector, estimated baseline mean for each interval.

convergence

an integer code, 0 indicates successful convergence, error code 1 indicates that the iteration limit maxIter had been reached.

iter

an integer value, number of interactions when stopped.

betaSE

a vector, estimated standard errors of beta.

betaVar

a matrix, estimated covariance matrix of beta.

baselineSE

a vector, estimated standard error of cumulative baseline mean.

See also